+++ tests/integration/create-channel.spec.holochain.js
@@ -1,0 +1,64 @@
+import 'regenerator-runtime/runtime.js'
+import { orchestrator } from './setup/tryorama'
+import { closeTestConductor, waitForState, findElementByText, findElementByClassandText, getElementProperty, beforeAllSetup } from './setup/helpers'
+import { TIMEOUT } from './setup/globals'
+
+orchestrator.registerScenario('New Message Scenario', async scenario => {
+  let aliceChat, page, closeServer
+  const callRegistry = {}
+  beforeAll(async () => {
+    const createPage = async() => await global.__BROWSER__.newPage();
+    // Note: passing in Puppeteer page function to instantiate pupeeteer and mock Browser Agent Actions
+    ({ aliceChat, page, closeServer } = await beforeAllSetup(scenario, createPage, callRegistry))  
+  }, TIMEOUT)
+
+  afterAll(async () => {
+    console.log("👉 Closing the UI server...")
+    await closeServer()
+    console.log("✅ Closed the UI server...")
+
+    console.log('👉 Shutting down tryorama player conductor(s)...')
+    await closeTestConductor(aliceChat, 'Create new Message')
+    console.log('✅ Closed tryorama player conductor(s)')
+  })
+
+  describe('New Channel Flow', () => {
+    it('creates and displays new message', async () => {
+      // *********
+      // register nickname
+      // *********
+      // verify page title
+      const pageTitle = await page.title()
+      expect(pageTitle).toBe('Elemental Chat')
+      // add agent nickname
+      const webUserNick = 'Bobbo'
+      await page.focus('.v-dialog')
+      await page.keyboard.type(webUserNick, { delay: 100 })
+      const [submitButton] = await findElementByText('button', 'Let\'s Go', page)
+      await submitButton.click()
+
+      // *********
+      // create channel
+      // *********
+      const newChannelTitle = 'Our Awesome New Room'
+      await page.type('#channel-name', newChannelTitle, { delay: 100 })
+      // press 'Enter' to submit
+      page.keyboard.press(String.fromCharCode(13))
+
+      // wait for create call response / load
+      const checkNewChannelState = () => callRegistry.createChannel
+      await waitForState(checkNewChannelState, 'done')
+      
+      // check for new channel title on page
+      const channels = await page.$eval('.channels-container', el => el.children);
+      expect(Object.keys(channels).length).toBe(1)
+      let newPage = page
+      const newChannelElement = await findElementByClassandText('div', 'v-list-item', newChannelTitle, newPage)
+      const newChannelHTML = await getElementProperty(newChannelElement, 'innerHTML')
+      expect(newChannelElement).toBeTruthy()
+      expect(newChannelHTML).toContain(newChannelTitle)
+    })
+  })
+})
+
+orchestrator.run()
