+++ tests/integration/setup/helpers.js
@@ -102,7 +102,6 @@ export const awaitZomeResult = async (
 
 /// Holo Test helpers:
 // -------------------
-
 export const holoAuthenticateUser = async (modalElement, email, password, type = 'signup') => {
   const createCredentialsLink = await findElementByText('a', 'Create credentials', modalElement)
   await createCredentialsLink.click()
@@ -134,69 +133,62 @@ export const holoAuthenticateUser = async (modalElement, email, password, type =
   return { email, password, confirmation }
 }
 
-
-
 /// Test Setup helpers:
 // -------------------
-
-export const beforeAllSetup = async scenario => {
-  console.log('👉 Setting up players on elemental chat...')
-    // Tryorama: instantiate player conductor
-    const [alice] = await scenario.players([conductorConfig], false)
-    await alice.startup()
-    // Tryorama: install elemental chat on both player conductors
-    const [[aliceChatHapp]] = await alice.installAgentsHapps([[{ hAppId: INSTALLED_APP_ID, dnas: [elChatDna] }]]);
-    // Tryorama: grab chat cell from list of happ cells to use as the 'player'
-    ([aliceChat] = aliceChatHapp.cells);
-
-    // Tryorama: alice declares self as chatter
-    await aliceChat.call('chat', 'refresh_chatter', null);
-
-    console.log('Confirming empty state at start')
-    let stats = await aliceChat.call('chat', 'stats', {category: "General"})
-    console.log('starting stats : ', stats)
-    expect(stats).toEqual(stats, {agents: 0, active: 0, channels: 0, messages: 0})
-
-    // locally spin up ui server only (not holo env)
-    console.log('👉 Spinning up UI server');
-    const { ports, close: closeServer } = httpServers()
-
-    // Puppeteer: use pupeeteer to mock Holo Hosted Agent Actions
-    page = await global.__BROWSER__.newPage()
-
-    page.once('domcontentloaded', () => console.info('✅ DOM is ready'))
-    page.once('load', () => console.info('✅ Page is loaded'))
-    page.once('close', () => console.info('✅ Page is closed'))
-    if (WEB_LOGGING) {
-      page.on('pageerror', error => console.error(`❌ ${error}`))
-      page.on('console', message => {
-        const consoleMessage = message.text();
-        console[message.type()](`ℹ️  ${consoleMessage}`)
-        const messageArray = consoleMessage.split(' ')
-        try {
-          // determine if message is a registered api call
-          if (parseInt(messageArray[0])) {
-            messageArray.shift()
-            const callDesc = messageArray.join(' ')
-            const callAction = messageArray.pop()
-            const isCallAction = callAction === 'start' || callAction === 'done'
-            if (isCallAction) {
-              if (messageArray.length > 1 && !callDesc.includes('zome')) return
-              const callName = messageArray[0]
-              // set the call with most current action state
-              callRegistry[callName] = callAction
-            }
-          }
-        } catch (error) {
-          // if error, do nothing - message is not a logged call
-          return
-        }
-      })
-    }
-
-    // Puppeteer: emulate avg desktop viewport
-    await page.setViewport({ width: 952, height: 968 })
-    await page.goto(`http://localhost:${ports.ui}/dist/index.html`)
+export const beforeAllSetup = async (scenario, createPage, callRegistry) => {
+  // Tryorama: instantiate player conductor
+  console.log('Settng up players on elemental chat...')
+  const [alice] = await scenario.players([conductorConfig], false)
+  await alice.startup()
   
-    return { aliceChat, page, closeServer }
+  // Tryorama: install elemental chat on both player conductors
+  const [[aliceChatHapp]] = await alice.installAgentsHapps([[{ hAppId: INSTALLED_APP_ID, dnas: [elChatDna] }]]);
+  // Tryorama: grab chat cell from list of happ cells to use as the 'player'
+  let aliceChat;
+  ([aliceChat] = aliceChatHapp.cells);
+
+  // Tryorama: alice declares self as chatter
+  await aliceChat.call('chat', 'refresh_chatter', null);
+
+  // locally spin up ui server only (not holo env)
+  console.log('👉 Spinning up UI server');
+  const { ports, close: closeServer } = httpServers()
+
+  const page = await createPage()
+
+  page.once('domcontentloaded', () => console.info('✅ DOM is ready'))
+  page.once('load', () => console.info('✅ Page is loaded'))
+  page.once('close', () => console.info('✅ Page is closed'))
+  if (WEB_LOGGING) {
+    page.on('pageerror', error => console.error(`❌ ${error}`))
+    page.on('console', message => {
+      const consoleMessage = message.text();
+      console[message.type()](`ℹ️  ${consoleMessage}`)
+      const messageArray = consoleMessage.split(' ')
+      try {
+        // determine if message is a registered api call
+        if (parseInt(messageArray[0])) {
+          messageArray.shift()
+          const callDesc = messageArray.join(' ')
+          const callAction = messageArray.pop()
+          const isCallAction = callAction === 'start' || callAction === 'done'
+          if (isCallAction) {
+            if (messageArray.length > 1 && !callDesc.includes('zome')) return
+            const callName = messageArray[0]
+            // set the call with most current action state
+            callRegistry[callName] = callAction
+          }
+        }
+      } catch (error) {
+        // if error, do nothing - message is not a logged call
+        return
+      }
+    })
+  }
+
+  // Puppeteer: emulate avg desktop viewport
+  await page.setViewport({ width: 952, height: 968 })
+  await page.goto(`http://localhost:${ports.ui}/dist/index.html`)
+
+  return { aliceChat, page, closeServer }
 }
